2.3.0-SNAPSHOT 1.实现ICacheManager新增接口； 2.完善实现IBigTable类逻辑； 3.优化若干类格式和注释等。
